Reorganizing - folders for main gui parts
