+++ BuildingThemes/BuildingThemes.csproj
@@ -65,29 +65,31 @@
     <Compile Include="Detour\ZoneBlockDetour.cs" />
     <Compile Include="Detour\PoliciesPanelDetour.cs" />
     <Compile Include="GUI\BuildingItem.cs" />
-    <Compile Include="GUI\PolicyPanelEnabler.cs" />
-    <Compile Include="GUI\BlacklistModeCheckboxContainer.cs" />
+    <Compile Include="GUI\ThemePolicies\UIPoliciesThemeTab.cs" />
+    <Compile Include="GUI\ThemePolicies\PolicyPanelEnabler.cs" />
+    <Compile Include="GUI\ThemePolicies\BlacklistModeCheckboxContainer.cs" />
     <Compile Include="GUI\PreviewRenderer.cs" />
-    <Compile Include="GUI\ThemeManagementCheckboxContainer.cs" />
-    <Compile Include="GUI\UIBuildingFilter.cs" />
-    <Compile Include="GUI\UIBuildingItem.cs" />
+    <Compile Include="GUI\ThemePolicies\ThemeManagementCheckboxContainer.cs" />
+    <Compile Include="GUI\ThemeManager\UIBuildingFilter.cs" />
+    <Compile Include="GUI\ThemeManager\UIBuildingItem.cs" />
     <Compile Include="GUI\UIFastList.cs" />
-    <Compile Include="GUI\UIBuildingInfo.cs" />
-    <Compile Include="GUI\UIThemeItem.cs" />
-    <Compile Include="GUI\UINewThemeModal.cs" />
-    <Compile Include="GUI\UIThemeManager.cs" />
+    <Compile Include="GUI\ThemeManager\UIBuildingInfo.cs" />
+    <Compile Include="GUI\ThemeManager\UIThemeItem.cs" />
+    <Compile Include="GUI\ThemeManager\UINewThemeModal.cs" />
+    <Compile Include="GUI\ThemeManager\UIThemeManager.cs" />
     <Compile Include="GUI\UITitleBar.cs" />
     <Compile Include="GUI\UIUtils.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Detour\RedirectionHelper.cs" />
     <Compile Include="SerializableDataExtension.cs" />
-    <Compile Include="GUI\ThemePolicyContainer.cs" />
+    <Compile Include="GUI\ThemePolicies\ThemePolicyContainer.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="BuildingThemes.xml">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>mkdir "%25LOCALAPPDATA%25\Colossal Order\Cities_Skylines\Addons\Mods\$(SolutionName)"
